+++ CHANGES.rst
@@ -8,6 +8,9 @@ New Features
 
 - Change parallelization framework for better performance and flexibility. [#3806]
 
+- Plugin-specific viewer tools (footprint, catalogs, line lists, line analysis) now have a
+  toggle button in the plugin UI to enable/disable a custom toolbar specific to that context. [#3812]
+
 Cubeviz
 ^^^^^^^